refactor(topic-draft): moved createTopicDraft() from OkrUnitServiceUsers to OkrTopicDraftService and adjusted tests
