+++ backend/burning-okr/burning-okr-logic/src/main/java/org/burningokr/service/okrUnit/departmentservices/OkrUnitServiceUsers.java
@@ -30,7 +30,6 @@ public class OkrUnitServiceUsers<T extends OkrChildUnit> implements OkrUnitServi
   protected final Logger logger = LoggerFactory.getLogger(OkrUnitServiceUsers.class);
   protected UnitRepository<T> unitRepository;
   protected ObjectiveRepository objectiveRepository;
-  protected OkrTopicDraftRepository topicDraftRepository;
   protected ActivityService activityService;
   ParentService parentService;
   private EntityCrawlerService entityCrawlerService;
@@ -40,13 +39,11 @@ public class OkrUnitServiceUsers<T extends OkrChildUnit> implements OkrUnitServi
       ParentService parentService,
       UnitRepository<T> unitRepository,
       ObjectiveRepository objectiveRepository,
-      OkrTopicDraftRepository topicDraftRepository,
       ActivityService activityService,
       EntityCrawlerService entityCrawlerService) {
     this.parentService = parentService;
     this.unitRepository = unitRepository;
     this.objectiveRepository = objectiveRepository;
-    this.topicDraftRepository = topicDraftRepository;
     this.activityService = activityService;
     this.entityCrawlerService = entityCrawlerService;
   }
@@ -91,19 +88,6 @@ public class OkrUnitServiceUsers<T extends OkrChildUnit> implements OkrUnitServi
   @Override
   public Objective createObjective(Long unitId, Objective objective, User user) {
     throw new UnauthorizedUserException("Service method not supported for current user role.");
-  }
-
-  @Override
-  public OkrTopicDraft createTopicDraft(OkrTopicDraft topicDraft, User user) {
-
-    topicDraft.setParentUnit(null);
-
-    topicDraft = topicDraftRepository.save(topicDraft);
-    logger.info("Created Topic Draft: " + topicDraft.getName());
-
-    activityService.createActivity(user, topicDraft, Action.CREATED);
-
-    return topicDraft;
   }
 
   void throwIfCycleForDepartmentIsClosed(OkrUnit okrUnitToCheck) {
