+++ backend/burning-okr/burning-okr-logic/src/test/java/org/burningokr/service/okr/OkrTopicDraftServiceTest.java
@@ -1,11 +1,15 @@
 package org.burningokr.service.okr;
 
 import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertNotNull;
+import static org.mockito.ArgumentMatchers.any;
+import static org.mockito.ArgumentMatchers.eq;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
 import java.time.LocalDate;
 import java.util.*;
+
 import org.burningokr.model.activity.Action;
 import org.burningokr.model.okr.okrTopicDraft.OkrTopicDraft;
 import org.burningokr.model.users.User;
@@ -94,5 +98,27 @@ public class OkrTopicDraftServiceTest {
     okrTopicDraftService.deleteTopicDraftById(okrTopicDraftId, user);
 
     verify(activityService).createActivity(user, this.okrTopicDraft, Action.DELETED);
+  }
+
+  @Test
+  public void createOkrDraft_expectIsSavedToDatabase() {
+    OkrTopicDraft topicDraft = new OkrTopicDraft();
+
+    when(okrTopicDraftRepository.save(any())).thenReturn(topicDraft);
+
+    okrTopicDraftService.createTopicDraft(topicDraft, user);
+
+    verify(okrTopicDraftRepository).save(any());
+  }
+
+  @Test
+  public void createOkrDraft_expectCreatesActivity() {
+    OkrTopicDraft topicDraft = new OkrTopicDraft();
+
+    when(okrTopicDraftRepository.save(any())).thenReturn(topicDraft);
+
+    okrTopicDraftService.createTopicDraft(topicDraft, user);
+
+    verify(activityService).createActivity(eq(user), any(), eq(Action.CREATED));
   }
 }
